+++ app.js
@@ -5,8 +5,6 @@ var mongoose = require('mongoose');
 var MongoStore = require('connect-mongo')(express);
 var settings = require('./settings');
 var flash = require('connect-flash');
-var multer  = require('multer');
-var fs = require('fs');
 
 //路由
 var authRoutes = require('./app/controllers/auth');
@@ -41,50 +39,7 @@ app.use(express.methodOverride());
 app.use(express.cookieParser());
 
 //文件上传配置
-app.use(multer({ dest: './public',
-                rename: function (fieldname) {
-                            return fieldname;
-                        },
-                changeDest: function(dest, req, res) {
-                            if(req.originalUrl === '/profile/uploadPortrait'){
-                                return dest + '/images/portrait'
-                            }else{
-                                return dest;
-                            }
-                        },
-                onFileUploadData: function (file, data, req, res) {
-
-                            console.log('in on file up load data');
-                            console.log(file.originalname);
-
-                            if(req.originalUrl === '/profile/uploadPortrait'){
-                                var ext = path.extname(file.originalname);
-                                ext = ext.toLowerCase();
-
-                                console.log('ext: ' + ext);
-
-                                //检查文件格式,若格式不正确，则删除文件，取消上传。若为正确格式的文件，同一转换成png格式。
-                                if(ext !== '.jpg' && ext !== '.jpeg' && ext !== '.png'){
-                                    if(fs.existsSync(path.join(__dirname,"public/images/portrait/",file.name))){
-                                        fs.unlinkSync(path.join(__dirname,"public/images/portrait/",file.name));
-                                    }
-                                }else{
-                                    var oldPath = path.join(__dirname,"public/images/portrait/",file.name);
-
-                                    //等待文件全部上传
-                                    while(!fs.existsSync(oldPath)){
-
-                                    }
-
-                                    var newFileName = path.basename(file.name,ext) + '.png';
-                                    var newPath = path.join(__dirname,"public/images/portrait/",newFileName);
-                                    fs.renameSync(oldPath,newPath);
-
-                                }
-
-                            }
-
-                        }}));
+require('./config/multerConfig')(app);
 
 app.use(express.session({
     secret: settings.cookieSecret,
